Improve Conv2D backward pass + implement dilation
With dilation, the backward pass (for gradient descent) of
a convolutional layer can be expressed itself as a convolution.
No special function for the backward pass is needed anymore
(which allows to only use forward function that is more optimized
giving a speedup factor of 2.4 in average).

Note that the Cython backward function hasn't been removed.

Finally, please allow me to express this:
FUCK YEAH, WE DID IT! :relieved: